requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
asn1crypto==1.2.0
astroplan==0.7
astropy==4.2
astropy-healpix==0.5
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
cachetools==4.1.0
certifi==2020.6.20
cffi==1.13.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
conda==4.8.3
conda-package-handling==1.6.0
cosmoplotian==0.0.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
dm-tree==0.1.5
entrypoints==0.3
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.27.2
h5py==2.10.0
horovod @ file:///global/homes/s/swowner/mustafa/horovod_build/horovod/dist/horovod-0.19.5-cp37-cp37m-linux_x86_64.whl
idna==2.8
importlib-metadata==1.7.0
importlib-resources==5.1.0
ipykernel==5.3.0
ipyparallel @ file:///tmp/build/80754af9/ipyparallel_1593440601605/work
ipython @ file:///tmp/build/80754af9/ipython_1593447368578/work
ipython-genutils==0.2.0
jedi @ file:///tmp/build/80754af9/jedi_1592841891421/work
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyter-server-proxy==1.5.0
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mpi4py==3.0.0
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///tmp/build/80754af9/pandas_1592841632039/work
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
pyerfa==1.7.1.1
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-nersc-customs==0.2.0
python3-nersc-modster==0.3.0
python3-nersc-sitecustomize==0.2.0
python3-nerscjson==0.9
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
reproject==0.7.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.46
ruamel.yaml.clib==0.2.0
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1592502882224/work
scipy @ file:///tmp/build/80754af9/scipy_1592930510235/work
Send2Trash==1.5.0
simpervisor==0.3
six==1.12.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.2.0
tensorflow-metadata==0.27.0
tensorflow-probability==0.10.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado==6.0.4
tqdm==4.36.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.24.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
